refactor: package structure for CameraBox component
